// apps/CFW.SupperAdmin/src/api/model/databaseTableViewModel.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * CFW.AppHost | v1
 * OpenAPI spec version: 1.0.0
 */
import type { DatabaseColumnViewModel } from './databaseColumnViewModel';

export interface DatabaseTableViewModel {
  name?: string;
  columns?: DatabaseColumnViewModel[];
  /** @nullable */
  schema?: string | null;
}